+++ backend/src/utils/llm.py
@@ -36,7 +36,7 @@ class LLMWrapper:
             model_name = model_name.replace('openai-', '')
             chosen_model = ChatOpenAI(model=model_name, **self.kwargs)
         elif 'anthropic' in model_name:
-            anthropic_token = self.user_repo.get_token(key=UserTokenKey.ANTHROPIC_API_KEY.name) or ANTHROPIC_API_KEY
+            anthropic_token = ANTHROPIC_API_KEY
             if not anthropic_token:
                 raise ValueError("Anthropic API key not found") 
             self.kwargs['api_key'] = anthropic_token
@@ -50,14 +50,14 @@ class LLMWrapper:
             model_name = model_name.replace('ollama-', '')
             chosen_model = ChatOllama(model=model_name, **self.kwargs)
         elif 'groq' in model_name:
-            groq_token = self.user_repo.get_token(key=UserTokenKey.GROQ_API_KEY.name) or GROQ_API_KEY
+            groq_token = GROQ_API_KEY
             if not groq_token:
                 raise ValueError("Groq API key not found")
             self.kwargs['api_key'] = groq_token
             model_name = model_name.replace('groq-', '')
             chosen_model = ChatGroq(model=model_name, **self.kwargs)
         elif 'google' in model_name:
-            gemini_token = self.user_repo.get_token(key=UserTokenKey.GEMINI_API_KEY.name) or GEMINI_API_KEY
+            gemini_token = GEMINI_API_KEY
             if not gemini_token:
                 raise ValueError("Gemini API key not found")
             self.kwargs['api_key'] = gemini_token
@@ -74,7 +74,7 @@ class LLMWrapper:
     def embedding_model(self):
         chosen_model = None
         if 'openai' in self.model_name:
-            openai_token = self.user_repo.get_token(key=UserTokenKey.OPENAI_API_KEY.name) or OPENAI_API_KEY
+            openai_token = OPENAI_API_KEY
             if not openai_token:
                 raise ValueError("OpenAI API key not found")
             self.kwargs['api_key'] = openai_token